requirements.txt
Flask
requests
Pillow
moviepy
pydub
playsound
importlib-resources